refactor(mcp): split redis.rs into server, keys, and structures submodules
